refactor: remove FloatType from generation tests

+++ honeycomb-core/src/utils/generation.rs
@@ -512,7 +512,6 @@ impl<T: CoordsFloat> GridBuilder<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::common::FloatType;
 
     #[test]
     fn build_nc_lpc_l() {
@@ -604,7 +603,7 @@ mod tests {
 
     #[test]
     fn square_cmap2_correctness() {
-        let cmap: CMap2<FloatType> = GridBuilder::unit_squares(2).build2().unwrap();
+        let cmap: CMap2<f64> = GridBuilder::unit_squares(2).build2().unwrap();
 
         // hardcoded because using a generic loop & dim would just mean
         // reusing the same pattern as the one used during construction
@@ -705,7 +704,7 @@ mod tests {
     #[allow(clippy::too_many_lines)]
     #[test]
     fn splitsquare_cmap2_correctness() {
-        let cmap: CMap2<FloatType> = GridBuilder::split_unit_squares(2).build2().unwrap();
+        let cmap: CMap2<f64> = GridBuilder::split_unit_squares(2).build2().unwrap();
 
         // hardcoded because using a generic loop & dim would just mean
         // reusing the same pattern as the one used during construction
